backend/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
# paho-mqtt 1.6.4 is not available on PyPI in some environments; use 1.6.1 which is available
paho-mqtt==1.6.1
python-dotenv==1.0.0